Basic DB and login running
Currently a user database is created to check for login authorization, a
user login page is created that checks for user details and tells
whether the user has logged in as a student or as a faculty. Currently
the only users and passwords it supports are in the userauth.txt file
which is created just to fill the database for the first time, once the
server is running its datastore is persistent.